requirements.txt
mysqlclient>=2.0.3,<2.1.0

django-mysql>=3.10.0,<3.11.0

Django>=3.1.4,<3.2.0

djangorestframework>=3.12.2,<3.13.0

flake8>=3.8.4,<3.9.0

gunicorn>=20.0.4,<20.1.0

martor>=1.5.8,<1.6.0

django-cors-headers>=3.7.0,<3.8.0

django-elasticsearch-dsl>=7.1.4, <7.2.0

django_elasticsearch_dsl_drf>=0.21, <1.0

elasticsearch>=7.0.0,<8.0.0
